+++ bikespace_dashboard/js/components/sidebar/data-panel/issue_chart.js
@@ -1,8 +1,14 @@
-import {Component} from '../main.js';
+import {Component} from '../../main.js';
 import {defaults} from './plot_utils.js';
-import {issue_attributes as ia} from '../issue_attributes.js';
+import {issue_attributes as ia} from '../../issue_attributes.js';
 
 class IssueChart extends Component {
+  /**
+   * Base class for graphs, map, etc. Registers component with shared_state.
+   * @param {string} parent JQuery selector for parent element
+   * @param {string} root_id tag id for root div
+   * @param {Object} shared_state
+   */
   constructor(parent, root_id, shared_state) {
     super(parent, root_id, shared_state);
