+++ bikespace_dashboard/js/components/sidebar/data-panel/summary_box.js
@@ -1,6 +1,12 @@
-import {Component} from '../main.js';
+import {Component} from '../../main.js';
 
 class SummaryBox extends Component {
+  /**
+   * Base class for graphs, map, etc. Registers component with shared_state.
+   * @param {string} parent JQuery selector for parent element
+   * @param {string} root_id tag id for root div
+   * @param {Object} shared_state
+   */
   constructor(parent, root_id, shared_state) {
     super(parent, root_id, shared_state);
     this.build();
